move PubdataPricing
