+++ core/lib/zksync_core/src/l1_gas_price/singleton.rs
@@ -8,8 +8,9 @@ use tokio::{
 use zksync_config::GasAdjusterConfig;
 use zksync_eth_client::clients::QueryClient;
 
-use super::gas_adjuster::PubdataPricing;
 use crate::l1_gas_price::GasAdjuster;
+
+use super::PubdataPricing;
 
 /// Special struct for creating a singleton of `GasAdjuster`.
 /// This is needed only for running the server.
